feature: database objects created/moved
created/updated the dataclasses for the db objects
moved to separate objected file
updated insert statements in other files
